refactor: move getType & getChar functions to helper.js
